Introduce org.astro module dependency
This step adds a second module named `org.astro` and updates the
`com.greetings` module declaration to _require_ a package _exported_ by
the `org.astro` module (the exported package is also named `org.astro`).

The modules are compiled separately. In order for `com.greetings` to
have access to the `org.astro` module's contents, the `javac` command is
given a `-modulepath` argument where the `org.astro` module and the
types in its exported packages can be resolved.

The main method in `com.greeting` has been updated as well, printing out
the name of a type referenced from the `org.astro` module. The output to
the console should now read:

    $ ./step.sh
    ...
    Greetings from org.astro's world!

http://openjdk.java.net/projects/jigsaw/quick-start#greetingsworld