Assemble com.greetings into executable modular jar
This step uses the `jar` tool to assemble our `com.greetings` module
into a _modular jar_. Note the use of some of the many new options for
`jar` in JDK 9, especially `--main-class`. There is no longer any need to
manually author an entry in a MANIFEST.MF file (or for a build system to
do so on your behalf); this is all taken care of in a first-class
    fashion when dealing with modular jars.

Once the modular jar has been created, `jar` is invoked once more, this
time with the new `--print-module-descriptor` option, which presents the
user with helpful module metadata (some of which will be explored in
subsequent steps):

    Name:
      com.greetings
    Requires:
      java.base [ MANDATED ]
    Main class:
      com.greetings.Main
    Conceals:
      com.greetings

Note also that the invocation of `java` also becomes simpler when
working with modular jars. `java` need only know about the directory
where modular jars reside (`lib` in this case), and the name of the
module to be "run" (`com.greetings` in this case).

Console output from the main method should again remain the same, as the
program itself still has not been modified in any way:

    $ ./step.sh
    ...
    Greetings from main!

See http://openjdk.java.net/projects/jigsaw/quick-start#packaging