Introduce service provider and bind successfully
This step introduces module `org.fastsocket` as a service provider for
the `com.socket.spi.NetworkSocketProvider` interface. It does not export
any packages.

With this service provider module in place, the `com.greetings` main
method should now be able to load and use an implementation of
`com.socket.NetworkSocket` and the output to the console should read as
follows:

    $ ./step.sh
    ...
    Greetings from class org.fastsocket.FastNetworkSocket!

See http://openjdk.java.net/projects/jigsaw/quick-start#services