Introduce com.greetings package and main class
This step adds a simple main class, compiles it with javac and runs it
with java in the usual fashion. No Jigsaw and no modules yet; that will
come in the next step.

When this step is run, it should print the following to the console:

    $ ./step.sh
    ...
    Greetings from main!

See http://openjdk.java.net/projects/jigsaw/quick-start#greeting